[ rename ] controller 폴더 정리
