+++ auth/src/main/java/com/linkeleven/msa/auth/presentation/controller/follow/FollowController.java
@@ -1,4 +1,4 @@
-package com.linkeleven.msa.auth.presentation.controller;
+package com.linkeleven.msa.auth.presentation.controller.follow;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
