+++ auth/src/main/java/com/linkeleven/msa/auth/presentation/controller/user/UserQueryController.java
@@ -1,4 +1,4 @@
-package com.linkeleven.msa.auth.presentation.controller;
+package com.linkeleven.msa.auth.presentation.controller.user;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
